+++ clients/es.d.ts
@@ -104,6 +104,14 @@ declare class ES extends Service {
    */
   deletePackage(callback?: (err: AWSError, data: ES.Types.DeletePackageResponse) => void): Request<ES.Types.DeletePackageResponse, AWSError>;
   /**
+   * Provides scheduled Auto-Tune action details for the Elasticsearch domain, such as Auto-Tune action type, description, severity, and scheduled date.
+   */
+  describeDomainAutoTunes(params: ES.Types.DescribeDomainAutoTunesRequest, callback?: (err: AWSError, data: ES.Types.DescribeDomainAutoTunesResponse) => void): Request<ES.Types.DescribeDomainAutoTunesResponse, AWSError>;
+  /**
+   * Provides scheduled Auto-Tune action details for the Elasticsearch domain, such as Auto-Tune action type, description, severity, and scheduled date.
+   */
+  describeDomainAutoTunes(callback?: (err: AWSError, data: ES.Types.DescribeDomainAutoTunesResponse) => void): Request<ES.Types.DescribeDomainAutoTunesResponse, AWSError>;
+  /**
    * Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.
    */
   describeElasticsearchDomain(params: ES.Types.DescribeElasticsearchDomainRequest, callback?: (err: AWSError, data: ES.Types.DescribeElasticsearchDomainResponse) => void): Request<ES.Types.DescribeElasticsearchDomainResponse, AWSError>;
@@ -430,6 +438,109 @@ declare namespace ES {
      */
     DomainPackageDetails?: DomainPackageDetails;
   }
+  export interface AutoTune {
+    /**
+     * Specifies Auto-Tune type. Valid value is SCHEDULED_ACTION. 
+     */
+    AutoTuneType?: AutoTuneType;
+    /**
+     * Specifies details of the Auto-Tune action. See the Developer Guide for more information. 
+     */
+    AutoTuneDetails?: AutoTuneDetails;
+  }
+  export type AutoTuneDate = Date;
+  export type AutoTuneDesiredState = "ENABLED"|"DISABLED"|string;
+  export interface AutoTuneDetails {
+    ScheduledAutoTuneDetails?: ScheduledAutoTuneDetails;
+  }
+  export type AutoTuneList = AutoTune[];
+  export interface AutoTuneMaintenanceSchedule {
+    /**
+     * Specifies timestamp at which Auto-Tune maintenance schedule start. 
+     */
+    StartAt?: StartAt;
+    /**
+     * Specifies maintenance schedule duration: duration value and duration unit. See the Developer Guide for more information.
+     */
+    Duration?: Duration;
+    /**
+     * Specifies cron expression for a recurring maintenance schedule. See the Developer Guide for more information.
+     */
+    CronExpressionForRecurrence?: String;
+  }
+  export type AutoTuneMaintenanceScheduleList = AutoTuneMaintenanceSchedule[];
+  export interface AutoTuneOptions {
+    /**
+     * Specifies the Auto-Tune desired state. Valid values are ENABLED, DISABLED. 
+     */
+    DesiredState?: AutoTuneDesiredState;
+    /**
+     * Specifies the rollback state while disabling Auto-Tune for the domain. Valid values are NO_ROLLBACK, DEFAULT_ROLLBACK. 
+     */
+    RollbackOnDisable?: RollbackOnDisable;
+    /**
+     * Specifies list of maitenance schedules. See the Developer Guide for more information.
+     */
+    MaintenanceSchedules?: AutoTuneMaintenanceScheduleList;
+  }
+  export interface AutoTuneOptionsInput {
+    /**
+     * Specifies the Auto-Tune desired state. Valid values are ENABLED, DISABLED. 
+     */
+    DesiredState?: AutoTuneDesiredState;
+    /**
+     * Specifies list of maitenance schedules. See the Developer Guide for more information.
+     */
+    MaintenanceSchedules?: AutoTuneMaintenanceScheduleList;
+  }
+  export interface AutoTuneOptionsOutput {
+    /**
+     * Specifies the AutoTuneState for the Elasticsearch domain.
+     */
+    State?: AutoTuneState;
+    /**
+     * Specifies the error message while enabling or disabling the Auto-Tune.
+     */
+    ErrorMessage?: String;
+  }
+  export interface AutoTuneOptionsStatus {
+    /**
+     *  Specifies Auto-Tune options for the specified Elasticsearch domain.
+     */
+    Options?: AutoTuneOptions;
+    /**
+     *  Specifies Status of the Auto-Tune options for the specified Elasticsearch domain.
+     */
+    Status?: AutoTuneStatus;
+  }
+  export type AutoTuneState = "ENABLED"|"DISABLED"|"ENABLE_IN_PROGRESS"|"DISABLE_IN_PROGRESS"|"DISABLED_AND_ROLLBACK_SCHEDULED"|"DISABLED_AND_ROLLBACK_IN_PROGRESS"|"DISABLED_AND_ROLLBACK_COMPLETE"|"DISABLED_AND_ROLLBACK_ERROR"|"ERROR"|string;
+  export interface AutoTuneStatus {
+    /**
+     * Timestamp which tells Auto-Tune options creation date .
+     */
+    CreationDate: UpdateTimestamp;
+    /**
+     * Timestamp which tells Auto-Tune options last updated time.
+     */
+    UpdateDate: UpdateTimestamp;
+    /**
+     * Specifies the Auto-Tune options latest version.
+     */
+    UpdateVersion?: UIntValue;
+    /**
+     * Specifies the AutoTuneState for the Elasticsearch domain.
+     */
+    State: AutoTuneState;
+    /**
+     * Specifies the error message while enabling or disabling the Auto-Tune options.
+     */
+    ErrorMessage?: String;
+    /**
+     * Indicates whether the Elasticsearch domain is being deleted.
+     */
+    PendingDeletion?: Boolean;
+  }
+  export type AutoTuneType = "SCHEDULED_ACTION"|string;
   export type BackendRole = string;
   export type Boolean = boolean;
   export interface CancelElasticsearchServiceSoftwareUpdateRequest {
@@ -540,6 +651,10 @@ declare namespace ES {
      * Specifies advanced security options.
      */
     AdvancedSecurityOptions?: AdvancedSecurityOptionsInput;
+    /**
+     * Specifies Auto-Tune options.
+     */
+    AutoTuneOptions?: AutoTuneOptionsInput;
   }
   export interface CreateElasticsearchDomainResponse {
     /**
@@ -660,6 +775,30 @@ declare namespace ES {
   }
   export type DeploymentCloseDateTimeStamp = Date;
   export type DeploymentStatus = "PENDING_UPDATE"|"IN_PROGRESS"|"COMPLETED"|"NOT_ELIGIBLE"|"ELIGIBLE"|string;
+  export interface DescribeDomainAutoTunesRequest {
+    /**
+     * Specifies the domain name for which you want Auto-Tune action details.
+     */
+    DomainName: DomainName;
+    /**
+     * Set this value to limit the number of results returned. If not specified, defaults to 100.
+     */
+    MaxResults?: MaxResults;
+    /**
+     * NextToken is sent in case the earlier API call results contain the NextToken. It is used for pagination.
+     */
+    NextToken?: NextToken;
+  }
+  export interface DescribeDomainAutoTunesResponse {
+    /**
+     * Specifies the list of setting adjustments that Auto-Tune has made to the domain. See the Developer Guide for more information.
+     */
+    AutoTunes?: AutoTuneList;
+    /**
+     * Specifies an identifier to allow retrieval of paginated results.
+     */
+    NextToken?: NextToken;
+  }
   export interface DescribeElasticsearchDomainConfigRequest {
     /**
      * The Elasticsearch domain that you want to get information about.
@@ -946,6 +1085,17 @@ declare namespace ES {
   export type DomainPackageDetailsList = DomainPackageDetails[];
   export type DomainPackageStatus = "ASSOCIATING"|"ASSOCIATION_FAILED"|"ACTIVE"|"DISSOCIATING"|"DISSOCIATION_FAILED"|string;
   export type Double = number;
+  export interface Duration {
+    /**
+     *  Integer to specify the value of a maintenance schedule duration. See the Developer Guide for more information.
+     */
+    Value?: DurationValue;
+    /**
+     * Specifies the unit of a maintenance schedule duration. Valid value is HOURS. See the Developer Guide for more information.
+     */
+    Unit?: TimeUnit;
+  }
+  export type DurationValue = number;
   export interface EBSOptions {
     /**
      * Specifies whether EBS-based storage is enabled.
@@ -1081,6 +1231,10 @@ declare namespace ES {
      * Specifies AdvancedSecurityOptions for the domain. 
      */
     AdvancedSecurityOptions?: AdvancedSecurityOptionsStatus;
+    /**
+     * Specifies AutoTuneOptions for the domain. 
+     */
+    AutoTuneOptions?: AutoTuneOptionsStatus;
   }
   export interface ElasticsearchDomainStatus {
     /**
@@ -1172,6 +1326,10 @@ declare namespace ES {
      * The current status of the Elasticsearch domain's advanced security options.
      */
     AdvancedSecurityOptions?: AdvancedSecurityOptions;
+    /**
+     * The current status of the Elasticsearch domain's Auto-Tune options.
+     */
+    AutoTuneOptions?: AutoTuneOptionsOutput;
   }
   export type ElasticsearchDomainStatusList = ElasticsearchDomainStatus[];
   export type ElasticsearchInstanceTypeList = ESPartitionInstanceType[];
@@ -1782,6 +1940,7 @@ declare namespace ES {
   export type ReservedElasticsearchInstanceOfferingList = ReservedElasticsearchInstanceOffering[];
   export type ReservedElasticsearchInstancePaymentOption = "ALL_UPFRONT"|"PARTIAL_UPFRONT"|"NO_UPFRONT"|string;
   export type RoleArn = string;
+  export type RollbackOnDisable = "NO_ROLLBACK"|"DEFAULT_ROLLBACK"|string;
   export type S3BucketName = string;
   export type S3Key = string;
   export type SAMLEntityId = string;
@@ -1848,6 +2007,27 @@ declare namespace ES {
      */
     SessionTimeoutMinutes?: IntegerClass;
   }
+  export type ScheduledAutoTuneActionType = "JVM_HEAP_SIZE_TUNING"|"JVM_YOUNG_GEN_TUNING"|string;
+  export type ScheduledAutoTuneDescription = string;
+  export interface ScheduledAutoTuneDetails {
+    /**
+     * Specifies timestamp for the Auto-Tune action scheduled for the domain. 
+     */
+    Date?: AutoTuneDate;
+    /**
+     * Specifies Auto-Tune action type. Valid values are JVM_HEAP_SIZE_TUNING and JVM_YOUNG_GEN_TUNING. 
+     */
+    ActionType?: ScheduledAutoTuneActionType;
+    /**
+     * Specifies Auto-Tune action description. 
+     */
+    Action?: ScheduledAutoTuneDescription;
+    /**
+     * Specifies Auto-Tune action severity. Valid values are LOW, MEDIUM and HIGH. 
+     */
+    Severity?: ScheduledAutoTuneSeverityType;
+  }
+  export type ScheduledAutoTuneSeverityType = "LOW"|"MEDIUM"|"HIGH"|string;
   export interface ServiceSoftwareOptions {
     /**
      * The current service software version that is present on the domain.
@@ -1899,6 +2079,7 @@ declare namespace ES {
      */
     Status: OptionStatus;
   }
+  export type StartAt = Date;
   export interface StartElasticsearchServiceSoftwareUpdateRequest {
     /**
      * The name of the domain that you want to update to the latest service software.
@@ -1950,6 +2131,7 @@ declare namespace ES {
   export type TagKey = string;
   export type TagList = Tag[];
   export type TagValue = string;
+  export type TimeUnit = "HOURS"|string;
   export type UIntValue = number;
   export interface UpdateElasticsearchDomainConfigRequest {
     /**
@@ -2004,6 +2186,10 @@ declare namespace ES {
      * Specifies the Encryption At Rest Options.
      */
     EncryptionAtRestOptions?: EncryptionAtRestOptions;
+    /**
+     * Specifies Auto-Tune options.
+     */
+    AutoTuneOptions?: AutoTuneOptions;
   }
   export interface UpdateElasticsearchDomainConfigResponse {
     /**